Move register allocator to tool algo parity path
